requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
dvc==3.30.1
python-multipart==1.0.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
click==8.1.7
requests==2.31.0
motor==3.3.1
pymongo==4.5.0
pydantic-settings==2.0.3
